refactor(core): split system_apply/mod.rs (#477)
Move the System apply-layer out of mod.rs: types.rs holds the shared data
types (ApplyReport, *Input, SystemInfo, CpuTemp), and stub_impl.rs holds
the Linux/WSL no-op apply functions — cfg-gated as a whole module,
mirroring the existing freebsd_impl. mod.rs is now declarations and
cfg-gated re-exports.

Pure relocation; the per-function cfg attributes collapse into the module
gate. cargo check + clippy --tests clean, 639 tests green.

Bump 5.97.15 -> 5.97.16.
