+++ aifw-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aifw-ui",
-  "version": "5.97.15",
+  "version": "5.97.16",
   "private": true,
   "scripts": {
     "dev": "next dev",
